refactor: Menu, Product 일급 컬렉션 구현
