+++ src/main/java/kitchenpos/application/MenuService.java
@@ -10,7 +10,7 @@ import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
 import kitchenpos.domain.MenuProducts;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.Products;
 import kitchenpos.dto.request.MenuCreateRequest;
 import kitchenpos.dto.response.MenuResponse;
 import kitchenpos.exception.KitchenException;
@@ -42,31 +42,14 @@ public class MenuService {
         MenuGroup menuGroup = menuGroupRepository.findById(request.getMenuGroupId())
             .orElseThrow(() -> new KitchenException("메뉴가 속한 메뉴 그룹이 존재하지 않습니다."));
 
-        List<Product> products = productRepository.findAllByIdIn(request.getProductIds());
-        List<MenuProduct> menuProductList = convertToMenuProducts(request, products);
+        Products products = new Products(productRepository.findAllByIdIn(request.getProductIds()));
+        List<MenuProduct> menuProductList = products.convertToMenuProducts(request);
 
         Menu savedMenu = menuRepository
             .save(new Menu(request.getName(), request.getPrice(), menuGroup));
         MenuProducts menuProducts = new MenuProducts(menuProductList, savedMenu);
         menuProductRepository.saveAll(menuProducts.getMenuProducts());
         return MenuResponse.of(savedMenu);
-    }
-
-    private List<MenuProduct> convertToMenuProducts(MenuCreateRequest request,
-        List<Product> products) {
-        return request.getMenuProducts().stream()
-            .map(menuProductRequest -> {
-                Product product = findProduct(products, menuProductRequest.getProductId());
-                Long quantity = menuProductRequest.getQuantity();
-                return new MenuProduct(product, quantity);
-            }).collect(Collectors.toList());
-    }
-
-    private Product findProduct(List<Product> products, Long productId) {
-        return products.stream()
-            .filter(product -> productId.equals(product.getId()))
-            .findAny()
-            .orElseThrow(() -> new KitchenException("메뉴에 속한 상품이 존재하지 않습니다."));
     }
 
     public List<MenuResponse> list() {
