+++ src/main/java/kitchenpos/application/OrderService.java
@@ -7,19 +7,17 @@ import kitchenpos.dao.MenuRepository;
 import kitchenpos.dao.OrderLineItemRepository;
 import kitchenpos.dao.OrderRepository;
 import kitchenpos.dao.OrderTableRepository;
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.Menus;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderLineItems;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.dto.request.OrderCreateRequest;
-import kitchenpos.dto.request.OrderLineItemCreateRequest;
 import kitchenpos.dto.request.OrderStatusChangeRequest;
 import kitchenpos.dto.response.OrderResponse;
 import kitchenpos.exception.KitchenException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.CollectionUtils;
 
 @Service
 public class OrderService {
@@ -46,37 +44,14 @@ public class OrderService {
         final OrderTable orderTable = orderTableRepository.findById(request.getOrderTableId())
             .orElseThrow(() -> new KitchenException("존재하지 않는 테이블에 배정되어 있습니다."));
 
-        List<Menu> menus = menuRepository.findAllByIdIn(request.getMenuIds());
-        List<OrderLineItem> orderLineItemList = convertToOrderLineItem(request, menus);
+        Menus menus = new Menus(menuRepository.findAllByIdIn(request.getMenuIds()));
+        List<OrderLineItem> orderLineItemList = menus.convertToOrderLineItem(request);
 
         Order savedOrder = orderRepository
             .save(new Order(orderTable, LocalDateTime.now()));
         OrderLineItems orderLineItems = new OrderLineItems(orderLineItemList, savedOrder);
         orderLineItemRepository.saveAll(orderLineItems.getOrderLineItems());
         return OrderResponse.of(savedOrder);
-    }
-
-    private List<OrderLineItem> convertToOrderLineItem(OrderCreateRequest request,
-        List<Menu> menus) {
-        List<OrderLineItemCreateRequest> orderLineItems = request.getOrderLineItems();
-
-        if (CollectionUtils.isEmpty(orderLineItems)) {
-            throw new KitchenException("주문 항목이 비어있습니다.");
-        }
-
-        return orderLineItems.stream()
-            .map(orderLineItemCreateRequest -> {
-                Menu menu = findMenu(menus, orderLineItemCreateRequest.getMenuId());
-                Long quantity = orderLineItemCreateRequest.getQuantity();
-                return new OrderLineItem(menu, quantity);
-            }).collect(Collectors.toList());
-    }
-
-    private Menu findMenu(List<Menu> menus, Long menuId) {
-        return menus.stream()
-            .filter(menu -> menuId.equals(menu.getId()))
-            .findAny()
-            .orElseThrow(() -> new KitchenException("존재하지 않는 메뉴가 포함되어 있습니다."));
     }
 
     public List<OrderResponse> list() {
